fix a bug where the highlight was highlighting the wrong item when resizing the window
To be honest I'm not 100% certain *why* this happens, but the function passed as event listener - and by extensions the variables used within - seem to become memoized.

When you open a page `activePathname` is initialized with an example value of `a`, when you navigate to `/b` it is properly updated to `b` and the relating effects to that change occur without a hiccup. However the event listeners retain `a` as a value for `activePathname` which makes the selection jump when recalculating the style.

A different solution should be to use `useCallback` and properly draw up the dependencies of the functions, but I'll ignore this for now.